move export/import test to its own class
so it doesn't interfere with other tests that require the original server setup
