+++ rest/rest-resources/src/test/java/org/apache/brooklyn/rest/resources/ServerResourceTest.java
@@ -106,100 +106,6 @@ public class ServerResourceTest extends BrooklynRestResourceTest {
     }
 
     @Test
-    public void testExportPersistedState() throws Exception {
-        BasicApplication app = manager.getEntityManager().createEntity(EntitySpec.create(BasicApplication.class));
-        Location loc = manager.getLocationManager().createLocation(LocationSpec.create(SshMachineLocation.class));
-        byte[] zip = client().path("/server/ha/persist/export").get(byte[].class);
-        
-        List<String> entryNames = listEntryNames(zip);
-        assertTrue(Iterables.tryFind(entryNames, StringPredicates.containsLiteral(app.getId())).isPresent(), "entries="+entryNames);
-        assertTrue(Iterables.tryFind(entryNames, StringPredicates.containsLiteral(loc.getId())).isPresent(), "entries="+entryNames);
-
-        Entities.unmanage(app);
-    }
-
-    @Test
-    public void testExportPersistedStateWithBundlesThenReimport() throws Exception {
-        // export seems to preserve install order probably to minimise conflicts, so use deferred start to get the problematic order
-        OsgiBundleInstallationResult r2 = ((ManagementContextInternal) manager).getOsgiManager().get().installDeferredStart(null,
-                () -> new ByteArrayInputStream(Strings.lines(
-                        "brooklyn.catalog:",
-                        "  bundle: b2",
-                        "  version: 1",
-                        "  id: b2",
-                        "  itemType: entity",
-                        "  item:",
-                        "    type: b1"
-                ).getBytes()),
-                false).get();
-
-        ((ManagementContextInternal) manager).getOsgiManager().get().install(
-                () -> new ByteArrayInputStream(Strings.lines(
-                        "brooklyn.catalog:",
-                        "  bundle: b1",
-                        "  version: 1",
-                        "  id: b1",
-                        "  itemType: entity",
-                        "  item:",
-                        "    type: org.apache.brooklyn.core.test.entity.TestEntity"
-                ).getBytes()),
-                BrooklynBomYamlCatalogBundleResolver.FORMAT, false, null).get();
-
-        r2.getDeferredStart().run();
-
-        String yaml = "services: [ { type: b2 } ]";
-        Response response = client().path("/applications")
-                .post(Entity.entity(yaml, "application/x-yaml"));
-        HttpAsserts.assertHealthyStatusCode(response.getStatus());
-
-        org.apache.brooklyn.api.entity.Entity b2 = Iterables.getOnlyElement(Iterables.getOnlyElement(manager.getApplications()).getChildren());
-        Asserts.assertInstanceOf(b2, TestEntity.class);
-
-        byte[] zip = client().path("/server/ha/persist/export").get(byte[].class);
-
-        // restart the server, so it has nothing, then try importing
-        destroyClass();
-
-        File mementoDir = Os.newTempDir(getClass());
-        manager = RebindTestUtils.managementContextBuilder(mementoDir, getClass().getClassLoader())
-                .persistPeriodMillis(Duration.ONE_MINUTE.toMilliseconds())
-                .haMode(HighAvailabilityMode.MASTER)
-                .forLive(true)
-                .enablePersistenceBackups(false)
-                .emptyCatalog(true)
-//                .properties(false)
-                .setOsgiEnablementAndReuse(useOsgi(), true)
-                .buildStarted();
-        new BrooklynCampPlatformLauncherNoServer()
-                .useManagementContext(manager)
-                .launch();
-
-        initClass();
-
-        Asserts.assertNull(manager.getTypeRegistry().get("b2"));
-        Asserts.assertSize(manager.getApplications(), 0);
-
-        Response importResponse = client().path("/server/ha/persist/import").post(Entity.entity(zip, MediaType.APPLICATION_OCTET_STREAM_TYPE));
-        HttpAsserts.assertHealthyStatusCode(importResponse.getStatus());
-
-        Asserts.assertNotNull(manager.getTypeRegistry().get("b1"));
-        Asserts.assertNotNull(manager.getTypeRegistry().get("b2"));
-        org.apache.brooklyn.api.entity.Entity b2b = Iterables.getOnlyElement(Iterables.getOnlyElement(manager.getApplications()).getChildren());
-        Asserts.assertInstanceOf(b2b, TestEntity.class);
-        Asserts.assertEquals(b2b.getId(), b2.getId());
-    }
-
-    private List<String> listEntryNames(byte[] zip) throws Exception {
-        List<String> result = new ArrayList<>();
-        ZipInputStream zipInputStream = new ZipInputStream(new ByteArrayInputStream(zip));
-        ZipEntry entry;
-        while ((entry = zipInputStream.getNextEntry()) != null) {
-            result.add(entry.getName());
-        }
-        return result;
-    }
-    
-    @Test
     public void testGetHighAvailability() throws Exception {
         // Note by default management context from super is started without HA enabled.
         // Therefore can only assert a minimal amount of stuff.
